+++ src/main/java/sort/SelectionSort.java
@@ -5,8 +5,8 @@ public class SelectionSort {
 	private int[] originalArray;
 	public int[] sortedArray;
 	private int arrayLength;
-	private int compareCount;
-	private int swapCount;
+	private long compareCount;
+	private long swapCount;
 
 	public SelectionSort(int[] originalDataSet) {
 		originalArray = new int[originalDataSet.length];
@@ -38,11 +38,11 @@ public class SelectionSort {
 		}
 	}
 
-	public int getCompareCount() {
+	public long getCompareCount() {
 		return compareCount;
 	}
 
-	public int getSwapCount() {
+	public long getSwapCount() {
 		return swapCount;
 	}
